% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civetwebR-package.R
\docType{package}
\name{civetwebR-package}
\alias{civetwebR}
\alias{civetwebR-package}
\title{civetwebR: Embedded HTTP server for R using CivetWeb}
\description{
Lightweight HTTP server for R based on the CivetWeb library, supporting
static file serving and handling of HTTP requests.
}
\author{
\strong{Maintainer}: Vedo Alagic \email{vedo.alagic@outlook.com}

Authors:
\itemize{
  \item Vedo Alagic \email{vedo.alagic@outlook.com}
  \item Troy Hernandez \email{troy@cornball.ai} (\href{https://orcid.org/0009-0005-4248-604X}{ORCID})
}

Other contributors:
\itemize{
  \item cornball.ai [copyright holder]
  \item The CivetWeb developers (CivetWeb, bundled in src/civetweb.c) [copyright holder]
  \item Sergey Lyubka (Mongoose, from which CivetWeb descends) [copyright holder]
  \item The Mbed TLS Contributors (Mbed TLS, bundled in src/mbedtls/) [copyright holder]
}

}
\keyword{internal}
