% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{.dispatch_request}
\alias{.dispatch_request}
\title{Dispatch an HTTP request}
\usage{
.dispatch_request(method, path, req = NULL)
}
\arguments{
\item{method}{HTTP method.}

\item{path}{Request path.}

\item{req}{Request object.}
}
\description{
Internal: called by driver loop
}
