% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{.get_handler}
\alias{.get_handler}
\title{Lookup an HTTP handler}
\usage{
.get_handler(method, path)
}
\arguments{
\item{method}{HTTP method.}

\item{path}{Request path.}
}
\description{
Lookup an HTTP handler
}
