% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{group}
\alias{group}
\title{Register handlers under a common path prefix}
\usage{
group(prefix, expr)
}
\arguments{
\item{prefix}{Character(1). Path prefix (e.g. "/api/v1").}

\item{expr}{Code that registers handlers.}
}
\value{
Invisibly returns TRUE.
}
\description{
Register handlers under a common path prefix
}
