% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{req_header}
\alias{req_header}
\title{Helper to extract a header from a request}
\usage{
req_header(req, key)
}
\arguments{
\item{req}{A cw_request object.}

\item{key}{Header name.}
}
\description{
Helper to extract a header from a request
}
