% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{req_query}
\alias{req_query}
\title{Helper to extract a query parameter from a request}
\usage{
req_query(req, key)
}
\arguments{
\item{req}{A cw_request object.}

\item{key}{Parameter name.}
}
\description{
Helper to extract a query parameter from a request
}
