% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{static_files}
\alias{static_files}
\title{Static file handler (serves files from a directory)}
\usage{
static_files(
  dir,
  prefix = "/",
  index = c("index.html"),
  cache_control = NULL,
  list_dirs = FALSE,
  template = NULL
)
}
\arguments{
\item{dir}{Character(1). Root directory to serve from.}

\item{prefix}{Character(1). URL prefix.}

\item{index}{Character vector. Index files.}

\item{cache_control}{NULL or character(1).}

\item{list_dirs}{Logical. Enable directory listing.}

\item{template}{Character(1) or NULL. Optional template path override.}
}
\description{
Static file handler (serves files from a directory)
}
